// src/models/veiculo.ts
export interface VeiculoModel {
  id: number;
  placa: string;
  renavam: string;
  tipoVeiculo: string;
  chassi: string;
  marca: string;
  modelo: string;
  anoFabricacao: number;
  cor: string;
  cambio: string;
  anoModelo: number;
  quilometragem: number;
  combustivel: string;
  statusVeiculo: string;
  statusOrigem: string;
}